Feat: 쿼리 전용 인터페이스 구현

+++ UserManagement/IServices/IQuery.cs
@@ -12,7 +12,7 @@ namespace UserManagement.IServices
 		List<T> GetAll();
 
 		// 특정 데이터 조회
-		T GetOne(int id);
+		T GetOne(long id);
 
 		// 데이터 추가
 		void Create(T entity);
@@ -21,6 +21,6 @@ namespace UserManagement.IServices
 		void Update(T entity);
 
 		// 데이터 삭제
-		void Delete(int id);
+		void Delete(long id);
 	}
 }
